refactor: rename userInfo
