+++ src/components/index.js
@@ -16,7 +16,7 @@ import {
 import { API } from "./API";
 import { Card } from "./Card";
 import { Section } from "./Section";
-import { UserInfo } from "./User-info";
+import { UserInfo } from "./UserInfo";
 import { FormValidator } from "./FormValidator";
 import { PopupWithImage } from "./PopupWithImage";
 import { PopupWithForm } from "./PopupWithForm";
